migration [nfc]: Pull our migrations into their own module
Past time for this; they've grown to much longer than the rest of
the module, and they have a rather different flavor.
